Fix warnings by switching 'int' to 'size_t' in a few places.

+++ src/str/charsplit.c
@@ -1,4 +1,4 @@
-#include <stdint.h>
+#include <stddef.h>
 #include <stdlib.h>
 #include <string.h>
 
@@ -7,18 +7,18 @@
 char **charsplit(char *input, char token)
 {
     char *str = input;
-    int len = strlen(str); 
+    size_t len = strlen(str);
     char *buffer = malloc(len);
 
     // Storage for pieces
-    int piece_idx = 0;
-    int piece_length = 0;
+    size_t piece_idx = 0;
+    size_t piece_length = 0;
     char **pieces = malloc(sizeof(char *) * ALI_SPLIT_MAX_PIECES);
-    for (int i = 0; i < ALI_SPLIT_MAX_PIECES; i++) {
+    for (size_t i = 0; i < ALI_SPLIT_MAX_PIECES; i++) {
         pieces[i] = NULL;
     }
 
-    for (int i = 0; i < len; i++) {
+    for (size_t i = 0; i < len; i++) {
         char c = *str;
         buffer[piece_length] = c;
